mail password

+++ be-datn/master/app/Http/Controllers/Auth/NewPasswordController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
-use App\Mail\ResetPassAlert;
+use App\Mail\ResetPasswordSuccess;
 use App\Models\password_resets;
 use App\Models\User;
 use Illuminate\Auth\Events\PasswordReset;
+use Illuminate\Auth\Notifications\ResetPassword;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
@@ -23,14 +24,14 @@ class NewPasswordController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\View\View
      */
-    public function create(Request $request,$token)
-    {    $check_token = password_resets::where('token','=',$token)->first();
-        if($check_token){
-        return view('auth.reset-password', ['token' => $token, 'email' => $request->email]);
-        }else{
-            return redirect(route('login'));
-        }
-    }
+    // public function create(Request $request,$token)
+    // {    $check_token = password_resets::where('token','=',$token)->first();
+    //     if($check_token){
+    //     return view('auth.reset-password', ['token' => $token, 'email' => $request->email]);
+    //     }else{
+    //         return redirect(route('login'));
+    //     }
+    // }
 
     /**
      * Handle an incoming new password request.
@@ -42,25 +43,37 @@ class NewPasswordController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'token' => 'required',
-            'password' => 'required|confirmed',
-            'password_confirmation' => 'required|same:password'
-        ],[
-            'password.required' => 'mật khẩu không được bỏ trống',
-            'password.confirmed' => 'Mật khẩu không trùng khớp',
-            'password_confirmation.required' => 'Vui lòng nhập lại mật khẩu'
-        ]);
-   
-       
-       
+        // $request->validate([
+        //     'token' => 'required',
+        //     'password' => 'required|confirmed',
+        //     'password_confirmation' => 'required|same:password'
+        // ],[
+        //     'password.required' => 'mật khẩu không được bỏ trống',
+        //     'password.confirmed' => 'Mật khẩu không trùng khớp',
+        //     'password_confirmation.required' => 'Vui lòng nhập lại mật khẩu'
+        // ]);
+        $check_token = password_resets::where('token','=',$request->token)->first();
+        if($check_token){
             $check_user = User::where('email','=',$request->email)->first();
             if($check_user){
                 $check_user->password = Hash::make($request->password);
                 $check_user->save();     
-                Mail::to($request->email)->send(new ResetPassAlert($check_user));
+                Mail::to($request->email)->send(new ResetPasswordSuccess($check_user));
                 password_resets::where('token','=',$request->token)->delete();
-                Auth::login($check_user);
+                return response()
+                ->json([
+                    "messeges"=>"đổi pass thành công",
+                    "status"=>true
+                ]);
+                // Auth::login($check_user);
+            }else{
+                return response()
+                ->json([
+                    "messeges"=>"đổi pass thất bại",
+                    "status"=>false
+                ]);
+                // return redirect(route('login'));
+            }
             };
         return redirect('/');
